+++ src/SpeedyWeather.jl
@@ -102,9 +102,11 @@ include("dynamics/orography.jl")
 include("physics/land_sea_mask.jl")
 
 # VARIABLES
+include("dynamics/tracers.jl")
 include("dynamics/particles.jl")
 include("dynamics/clock.jl")
 include("dynamics/prognostic_variables.jl")
+include("dynamics/set.jl")
 include("physics/define_column.jl")
 include("dynamics/diagnostic_variables.jl")