+++ src/models/primitive_wet.jl
@@ -17,13 +17,13 @@ $(TYPEDFIELDS)"""
     AT,     # <:AbstractAtmosphere,
     CO,     # <:AbstractCoriolis,
     GO,     # <:AbstractGeopotential,
-    OR,     # <:AbstractOrography,
     AC,     # <:AbstractAdiabaticConversion,
     PA,     # <:AbstractParticleAdvection,
     IC,     # <:AbstractInitialConditions,
-    RP,     # <:AbstractRandomProcess,
     FR,     # <:AbstractForcing,
     DR,     # <:AbstractDrag,
+    RP,     # <:AbstractRandomProcess,
+    OR,     # <:AbstractOrography,
     LS,     # <:AbstractLandSeaMask,
     OC,     # <:AbstractOcean,
     LA,     # <:AbstractLand,
@@ -68,9 +68,12 @@ $(TYPEDFIELDS)"""
     adiabatic_conversion::AC = AdiabaticConversion(spectral_grid)
     particle_advection::PA = NoParticleAdvection()
     initial_conditions::IC = InitialConditions(PrimitiveWet)
-    random_process::RP = NoRandomProcess()
     forcing::FR = NoForcing()
-    drag::DR = NoDrag()
+    drag::DR = NoDrag()  
+  
+    # VARIABLES
+    random_process::RP = NoRandomProcess()
+    tracers::TRACER_DICT = TRACER_DICT()
     
     # BOUNDARY CONDITIONS
     orography::OR = EarthOrography(spectral_grid)
@@ -97,7 +100,7 @@ $(TYPEDFIELDS)"""
     optical_depth::OD = ZeroOpticalDepth(spectral_grid)
     shortwave_radiation::SW = NoShortwave(spectral_grid)
     longwave_radiation::LW = JeevanjeeRadiation(spectral_grid)
-    stochastic_physics::SP = StochasticallyPerturbedParameterizationTendencies(spectral_grid)
+    stochastic_physics::SP = NoStochasticPhysics()
     
     # NUMERICS
     time_stepping::TS = Leapfrog(spectral_grid)