+++ src/output/netcdf_output.jl
@@ -1095,6 +1095,39 @@ function output!(
     return nothing
 end
 
+"""Defines netCDF output for a specific variables, see `VorticityOutput` for details.
+Fields are $(TYPEDFIELDS)"""
+@kwdef mutable struct TracerOutput <: AbstractOutputVariable
+    name::String = "tracer1"
+    unit::String = "?"
+    long_name::String = "tracer1"
+    dims_xyzt::NTuple{4, Bool} = (true, true, true, true)
+    missing_value::Float64 = NaN
+    compression_level::Int = 3
+    shuffle::Bool = true
+    keepbits::Int = 15
+end
+
+"""$(TYPEDSIGNATURES)
+`output!` method for `variable`, see `output!(::NetCDFOutput, ::VorticityOutput, ...)` for details."""
+function output!(
+    output::NetCDFOutput,
+    variable::TracerOutput,
+    progn::PrognosticVariables,
+    diagn::DiagnosticVariables,
+    model::AbstractModel,
+)
+    tracer = output.grid3D
+    tracer_grid = diagn.grid.tracers_grid[Symbol(variable.name)]
+    RingGrids.interpolate!(tracer, tracer_grid, output.interpolator)
+
+    round!(tracer, variable.keepbits)
+    i = output.output_counter   # output time step to write
+    output.netcdf_file[variable.name][:, :, :, i] = tracer
+    return nothing
+end
+
+
 """
 $(TYPEDSIGNATURES)
 Checks existing `run_????` folders in `path` to determine a 4-digit `id` number